refactor : LottoService, LottoController, StringUtil 정리
